Remove bench modifiers pending randomized bench
This module implements some extremely basic caching. Some of the other
frameworks also implement some level of caching. The practical effect
of this on synthetic benchmarks is that real-world scenarios of
often differing requests and responses are not reflected. Instead I
need to randomize the bench request data in a way that reflects real
world usage. Until then I'm not going to disable caching in scrud since
other frameworks also benefit from it.