Remove crappy send cache, no gzip on lob bench
None of the other libs appear to be auto-gzipping so the use of gzip
in the lob benchmarks leads to a massive perf hit in comparison. I
could add the related modules and what not to ensure the others are
gzipping, but that just adds external deps that may not reflect real
world performance for all. So, no gzip accepting headers will be set in
the banchmarks.